backend/requirements-dev.txt
# Development and Testing Dependencies for FastVue Backend
# Install with: pip install -r requirements-dev.txt

# Include production dependencies
-r requirements.txt

# Testing Framework
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-xdist>=3.5.0  # Parallel test execution

# Test Data Generation
factory-boy>=3.3.0
faker>=22.0.0

# Property-Based Testing
hypothesis>=6.92.0

# HTTP Testing (already in requirements.txt, but ensure version)
# httpx>=0.26.0

# Time Mocking
freezegun>=1.4.0

# Coverage
coverage[toml]>=7.4.0

# Code Quality
ruff>=0.1.0
black>=24.0.0
mypy>=1.8.0

# Type stubs
types-redis>=4.6.0
types-python-dateutil>=2.8.0
